Small cleanups
